Fixed decoding of maildata
Previously fetching from imap source lead to an error, as the fetched data was uft-6 encoded and not decoded to be useable as a string. Added additional function decode_bytes_to_str to fix this. I then only fixed the IMAPSource class, one may need to decode data from other sources if required.